requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.3
django-import-export==3.3.1
redis==5.0.1
celery==5.3.4
python-decouple==3.8
gunicorn==21.2.0
whitenoise==6.6.0
setuptools==80.9.0
modeltranslation==0.25

djangorestframework-simplejwt==5.3.0
drf-spectacular==0.27.0
django-rest-auth==0.9.5
django-allauth==0.57.0
drf-nested-routers==0.93.4
django-rest-passwordreset==1.3.0
django-redis==6.0.0
pillow==12.0.0